# use goroutines to download images (4x faster)