use [goroutines](https://gobyexample.com/goroutines) to speed up download times for big galleries

Downloading a 10 image gallery:
| goroutine | no goroutine |
| :------------: | :-----------: |
| 422.3982ms | 6.4326704s | 
| 306.2332ms | 1.4970517s |
| 265.9872ms | 1.6601041s |
| 517.3667ms | 1.62706s     |
| 321.5757ms | 1.6395222s |

doesn't make much of a difference for single image downloads, but for large galleries, the difference is very noticable.