@cmd777 2023-01-09
> i tried out this PR and i noticed you didn't add any waitgroup in [downloader_nos](https://github.com/noornee/reddit-dl/blob/3a7f3d3a8035eda6f14d3afcdc935cc1a380598d/service/downloader.go#L79)
> i tried adding a wait group and ran the script but there wasnt any notable difference and then it dawned on me that you added the waitgroup to the [downloader](https://github.com/noornee/reddit-dl/blob/3a7f3d3a8035eda6f14d3afcdc935cc1a380598d/service/downloader.go#L50) function because it's downloading 2 different files(audio & video) hence concurrency!...TIL :)

Spot on. It also concurrently downloads all images from the media_url in the [main](https://github.com/noornee/reddit-dl/pull/6/commits/f0e699bd92f9fec56fa85914bd321438e28ba367#diff-38035a4d07316d9007b9d9ad9016d46f0f25321f17a3fef07c2b4159d8360528) program, (prob the biggest speedup tho, but again, not much of a difference for 1-2 images)

Also, perhaps instead of requiring go to install the tool, how about a workflow that builds the main package and then uploads the files?