@noornee 2023-01-09
> Spot on. It also concurrently downloads all images from the media_url in the [main](https://github.com/noornee/reddit-dl/pull/6/commits/f0e699bd92f9fec56fa85914bd321438e28ba367#diff-38035a4d07316d9007b9d9ad9016d46f0f25321f17a3fef07c2b4159d8360528) program, (prob the biggest speedup tho, but again, not much of a difference for 1-2 images)
> 

yeah, i noticed that!
> Also, perhaps instead of requiring go to install the tool, how about a workflow that builds the main package and then uploads the files?

hmm, i haven't done this before
I'm gonna research how it's done and see how that works andd how are users going to install it ?