@cmd777 2023-01-09
> hmm, i haven't done this before
> I'm gonna research how it's done and see how that works andd how are users going to install it ?

They can install it with go install, or if you do make a workflow for it, they will be able to just go to the releases page and download it there, though I'm not sure if it interferes with go get / go install (I don't think it will, though.)

Feel free to use [these](https://github.com/cmd777/lex/tree/main/.github/workflows) workflows as a template, and edit it to your liking, or if you need help, then I'll try to help ya with those workflows, but no promises, i'm not big into workflows.