@noornee 2023-01-09
> or if you do make a workflow for it, they will be able to just go to the releases page and download it there

i see, so they wouldn't need to have go installed to be able to run it ? neat!

> Feel free to use [these](https://github.com/cmd777/lex/tree/main/.github/workflows) workflows as a template, and edit it to your liking (probably should change the cron job, and remove minify & zipping)

for some reason, i feel you're my guardian angel!!


> if you need help, then I'll try to help ya with those workflows, but no promises, i'm not big into workflows.

Thankss, I'm going to try doing this myself and if i have any issues with it, i shall come for ya!