@cmd777 2023-01-09
> i see, so they wouldn't need to have go installed to be able to run it ? neat!

Yeah, exactly. 

> for some reason, i feel you're my guardian angel!!

:)

> Thankss, I'm going to try doing this myself and if i have any issues with it, i shall come for ya!

Alright, best of luck with it